Figured out issue with tank going to white flag
The robot tank was going to the white flag again because of the if statement 'if (teamGame && getFlag() == Flags::Null)'. The robot tank starts the game holding its own flag so the second if condition would resolve to false, and the code would go to the else if statement which wasn't fully implemented (though I'm not sure why it went to a white flag. It should have gone to it's own base. Anyway it looks like the robot tank sort of goes back to its base. Like it goes halfway and then gets stuck.