Small unexpectedly needed addition
Added code to make robot tank drop any white flags it picks up. Still have issue with tank not going all the way to its own base. It looks like it gets stuck exactly in the middle of the map. The coordinates printed out are 0,190,0. Also looking over the code given for getting the base coordinates, in the World.h file, the getBase() function looks incorrect. Like it tries to return a variable p from a TeamBases object, but in the TeamBases.h file it shows no variable 'p'.