Robot returns to its own base after grabbing flag
The issue stemmed from the "v[0] = flagPos[0] - position[0];" and the "v[1] = flagPos[1] - position[1];" lines of code in the final section of the do update motion. Because I wasn't using or setting the flagPos array in the elseif portion of my code, it was likely going to the middle of the map as that array was probably being set to 0,0,0.