Actually now it returns to base.
I had some testing variables of sort when I was trying to figure out whether the base coordinates were correct or if the robot itself was getting lost, were I basically set the return base as RedTeam. And it must of been on that particular run where everything worked the enemy was the red team. I tried running it again and saw that it didn't now and realized the mistake.