Minor success
So the enemy robot tank finally follows my tank and stays by it and follows it around. The issue was that my own tank (the localplayer running the bzfs executable) doesn't come up when using the World::getWorld->getPlayer() function. I had read the assignment instructions wrong and had thought the instructions said that the robot tank wouldn't get its OWN coordinates from the function. What was happening originally was the tank was taking its coordinates and adding it to itself.