More functionality added for Ally robot tank
The large if/else statement was updated so that the Ally robot tank may either seek out and return enemy flags or follow the LocalPlayer tank depending on the distance between them. I encountered an issue while trying to make a set distance threshold for which the Robot would either decide to follow the leader or seek a flag. The issue was after figuring out the center of mass point, I was using that x,y coordinate with the hypotf() function to calculate the distance, but that was wrong and giving me a far larger distance than what really was. This is because the hypotf() function calculates the hypotenuse of those points from (0,0), not from where the robot tank actually is. To get a correct center of mass distance, I took the difference between the Robot's position and the center of mass, and used that in the hypotf() function (note that this was what was done in the already existing example but I missed it/forgot about it).

Also note that there was an error with updating between the branches. The most recent master branch update was supposed to have been for this branch.