Changed separation mechanism
The assignment wanted us to implement separation, which I already had done in a different way not knowing it would ask for it later. It seems to be working but I would like to go back and rename some variables to make the code look more clear. Right now some of the arrays are just single letters.