Changed variables and fixed unseen error
For the separation branch of the if/else statement, when calculating the strength, the threshold was having an incorrect c[2] subtracted from it, where c[2] was just the difference between the localplayer z coordinate and the robot z coordinate. It should have been the distance between the local player and the robot, like was being done for comparisons against other robots. Also changed the variables to something more readable.