Files compile and run, though does nothing really
First issue was that the after adding in the two new ARegion files, the makefile wouldn't work. It appeared to be working fine at first when the program was using just the ARegion.h file and not calling to any of the details in the ARegion.cxx file, but when we tried to use the updated constructor it kept giving the error that it was undefined. Looking at the files, we could see that the .o version of the ARegion.cxx file wasn't being created. Making modifications to the Makefile in the /src/bzflag folder eventually got it to compile.

Second issue was using the vectors. The insert() function seems to only work (as in the program will only compile) if you feed it a iterator as the first parameter. Lots of time was wasted trying to feed it an integer and varying forms of the object to be inserted.

Another issue was after changing the insert() parameter to a iterator, we kept getting segmentation faults. Eventually we just changed this to use the push_back() function and it seems to be working alright, as in it compiles and runs but doesn't really do anything useful in terms of the assignment.